Extracted setting to a separate settings.ts
